requirements.txt
git+https://github.com/Rapptz/discord.py.git
psutil>=5.9.0
requests>=2.32.4
aiohttp>=3.12.15
beautifulsoup4>=4.13.4
Pillow>=9.0.0
